React Reorganize - components, second level
