+++ App/Containers/Components/ApproveFriendRow.js
