+++ App/Containers/Components/FriendListItem.js
