+++ App/Containers/Components/FriendTeamTile.js
