+++ App/Containers/Components/FriendTile.js
