+++ App/Containers/Components/PendingFriendRow.js
