+++ App/Containers/Components/Styles/ApproveFriendRowStyles.js
