+++ App/Containers/Components/Styles/FriendListItemStyles.js
