+++ App/Containers/Components/Styles/FriendTeamTileStyles.js
