+++ App/Containers/Components/Styles/FriendTileStyles.js
