+++ App/Containers/Components/Styles/FriendsStyles.js
