+++ App/Containers/Components/Styles/PendingFriendRowStyles.js
