+++ App/Containers/Components/Styles/SearchBarStyles.js
