+++ App/Containers/Components/TeamListRow.js
