+++ App/Containers/Stories/StoryScreen.js
@@ -3,7 +3,8 @@ import { ListView, View, Text } from 'react-native'
 import firebaseApp from '../Firebase'
 import * as firebase from 'firebase'
 import { connect } from 'react-redux'
-import { fetchJourney } from '../Redux/StoriesRedux'
+import { fetchJourney } from '../../Redux/StoriesRedux'
+
 import StoryListItem from '../Components/StoryListItem'
 import SearchBar from '../Components/SearchBar'
 import styles from './Styles/StoryScreenStyles'
