+++ App/Containers/UserProfile/ViewFriends.js
